Create countInFile.java
Created countInFile.java class, this class will search in the text file the number of instances the letters to search for occur.  This class has 3 methods, one for setting the field values, another for searching, and the last method for returning the results.  

This class is still in the process of being created.