Update converter.java
Added the convert and getResult methods into the class.  The convert method is responsible for calling the correct type of measurement conversion class, and assigning the value of output to the conversion result returned value.  The getResult method returns the value of the result of the conversion, which is represented by the variable output.